Use current UK time when calculating embargo datetimes
Before: the embargo datetime expiry was based on the current server
datetime. For most of our servers, this is GMT, which means that e.g. in
British Summer Time, the time the user selected, based on their own
timezone would not be the time the embargo actually expired.

Now: we explicitly use the London timezone, and mark the input field as
such for clarity.